Allow application to set custom jline terminal
sbt needs to be able to override the jline Terminal that instance that
is passed into the console line reader. There are two reasons for this:

1. sbt recently upgraded to jline3 and
   TerminalBuilder.builder().(...).build() will create a system
   terminal that interferes with the sbt line reader
2. sbt now supports relaying io to a remote client and this requires
   using a custom instance of jline Terminal that performs this
   forwarding. Without using the custom terminal, the remote client
   is unable to send input or receive output which makes the console
   task useless.

To make this work, we add a system property that gives the name of a
scala object that contains a no argument method called
getJLine3Terminal. When that property is set, we can try to load the
object using reflection and invoke the method. If the property is unset
or reflection fails, we can just fall back to the default builder.

Here is an example of the 2.13 console working with the sbt thin client
after this change: https://asciinema.org/a/K9uxsnISyzjgonxv3P6gPwjr2