MultiStringSetting preserves remaining args
Previously, the setting reversed remaining input args.
This is so that later settings override earlier settings.
But inputs may be due to colon `-Setting:value` or `-Setting value`.
The latter form presents all remaining args, so reversing those
inputs is not appropriate. Instead, reverse the result.

For a given usage of `-Wconf`, configuration order is preserved.
A subsequent usage of `-Wconf` is prepended to the configuration,
so that last usage wins.

`-Wconf any:s cat=lint:ws -Wconf cat=lint:ws any:s`

permits the lint summary because the second usage has precedence
and the ws is first in the configuration.

`-Wconf cat=lint:ws any:s -Wconf any:s cat=lint:ws`

suppresses the summary because the s is first in the second `-Wconf`.