Don't emit read of omitted `private[this] var` field
This is not trivial to fix unfortunately.

The `constructors` phase creates the primary constructor and moves
statements from the class body into it. While doing that, it rewrites
reads of to class parameter fields to the constructor parameters, except
for variables - otherwise writes to the variable are not reflected
(this is the fix done in https://github.com/scala/scala/pull/7688).

Only after that, the phase checks which `private[this]` fields are not
used. The implementation for that uses the triage of the class body
elements (primary constr, aux constrs, other defs) done in the first
step.

So if we remove the `private[this] var` field in the second step
(because it's never written, and never read outside of the constructor),
the constructor ends up with an access to a non-existing field.

The fix here delays rewriting reads of class parameter fields to
constructor parameters until after the omittable fields are computed.