Fix regression in LocalVariableTable attribute
The change to fix an off-by-one-statement error in the start
BCI of entries in the LocalVariableTable caused the on-by-default
version of dead code elimination to incorrectly remove entries
in some cases.

This commit changes `removeUnusedLocalVariableNodes` to do a full
back up one instruction to include the initial xSTORE in the range of
instructions in the search is performed.