Fix exhaustivity uncheckableType's logic for tuples
As tuples (e.g. Tuple2) are case classes, for which `enumerateSubtypes`
returns `List(List(tp))` therefore making it "checkable", any tuple type
was accidentally always passing (or at least it has since that `isCase`
branch was added to `enumerateSubtypes`).  This was leading to false
positives in the exhaustivity checking, as a tuple of non-sealed types
was being deemed inexhaustive with a wildcard `(_, _)` counter-example.