Predef.unit to workaround "a pure expression does nothing"
For some reason, in the recent releases "a pure expression does nothing in statement position" has gotten more accurate. Maybe because it moved to RefCheck in #8995?
Regardless, it's sometimes tricky to write `build.sbt` because often I would need to invoke multiple tasks such as `compile.value` and `copyResources.value` without using the return value, and it would cause "a pure expression does nothing in statement position" error.

As a workaround I propose `Predef.unit(a: Any): Unit = ()`, which forces the evaluation but discards it and returns the unit value, which the compiler seems to be happy with.