Set the tpt of extension methods' DefDef explicitly
The default `tpt` is computed from the method symbol
by mapping with`AsSeenFromMap` which extends `KeepOnlyTypeConstraints`.
This drops the `@uncheckedVariance` annotation,
which in turn alters the bound of a type parameter,
which in turn causes the creation of a new type parameter,
which in turn alters the method result type.