Check exhaustivity of case class unapplySeq extractors too
ProductExtractorTreeMaker handles non-custom extractors, i.e. the
synthetic ones generates for case classes.  I think this case was just
accidentally missing from the original commit.

The change to run/virtpatmat_unapplyprod makes sense to me - it's an
inexhaustive match, albeit with complicated counter-examples... :(

This is the actually the last remaining tree maker, so none will back
off after this, from my reading - but I kept the back-off stuff as is,
keeping the diff minimal, to try to land this as easily as possible.