Fix match error in DelegatingReporter.scala
8b2db7a792f3dbf47d31d6e543b353b4e1a42834 introduced a regression because
it is not always the case that the source passed in to
DelegateReporter.makePosition is an instance of AbstractZincFile.
53bc41acc7d878617f0d8b6f9400d606e8191637 partially fixed this by
handling PlainFile but the proper fix is to handle AbstractFile which is
what was done prior to 8b2db7a792f3dbf47d31d6e543b353b4e1a42834

This was detected by running the community build in the shapeless
project.

Rewritten from sbt/zinc@a05e58ab8511b6171346d0b5fb98c76fe8722e41