Avoid swallowing error message and crashing for X#Inaccessible
The extra ref-checking in d7814a23 (aka TypeTreeWithDeferredRefCheck)
seems to have been added in a way that obscures reporting of access
errors. This leads to downstream crashes in later phases.

This failure used to be caught by a backstop in the pickler phase
which emitted a "erroneous or inaccessible type" error, which regressed
in cf477a0e40 but was fixed in the commit prior to this one.