Patmat: don't drop abstract subtypes prematurely
In the code:

    sealed trait A
    sealed trait AB extends A
    sealed trait AA extends A

    object Lib {
      def f(a: A): Boolean = a match {
           case _:AA => true
           case _:AB => false
      }
    }

the reason the compiler was giving an unreachable warning ("warning:
unreachable code case _:AB => false") was that A, AB and AA are all
uninhabited due to the fact that they're both abstract and sealed, i.e.
they have no subclasses.  Which means that `case _:AB =>` is indeed
unreachable, as is, in fact, `case _:AA => ` (the reason it reports the
`case _: AB` branch I think is because the pattern match analyser is
investigating "if the type is AA, are any of the branches unreachable?"
first).

The fact that in the REPL you can extend a sealed trait is kind of an
implementation detail, a convenience for REPL usage.  I almost wrote the
test as a REPL test to reproduce it, but then remembered I could use
Java instead.

Just to get it out of the way: suppressing the warning when in the REPL
would be a terrible idea, IMO.

Let's just not filter sealed, abstract subclasses.