Fix async phase under toolbox compilation
The async phase has an optimization to skip units if they are known not
to have any async calls in them. This is done by storing the `SourceFile`
of every tree that is handed to `markForAsyncTransform`, and then skipping
over any unit that does not have such a source file.

However, toolbox parsing returns a tree with positions in a synthetic file
named "<toolbox>"; however, the unit passed to the compiler later on in
`compile` is made without a source file; therefore, the unit is skipped by
the async phase, and the `@compileTimeOnly` errors are emitted.

As a cheap fix, this commit always transforms units with `NoSourceFile`.