Keep `private[this] var` class parameter fields that are only read
Reads of `var` class parameters in the constructor are not rewritten to
the constructor parameter, they keep referring to the field to ensure
potential field writes are reflected (this was fixed in #7688).

This means that the corresponding field cannot be elided even if it's
never written.

The alternative solution would be to detect `private[this] var` fields
that are only read in the constructor. However this is difficult to
implement, see #9143.