Fix the List rewrite in patmat to use 2.13's UnapplySeqWrapper
I didn't double-check, but what I expect happened is that all the
pattern matches that included a "case List()" stopped producing any
exhaustivity/unreachable warnings.  And then my guess is the impact
wasn't understood and the checkfiles were just updated?

In the next commit strict pattern matching becomes the default, which,
when bootstrapping, found a pattern match in the compiler that uses
"case List()" (with -Werror enabled).