Block opaque type aliases
Following a discussion, it makes sense to prevent opaque type
aliases. When the alias is visible in Scala 2 then the
intended API of the opaque type is broken.

It is possible to re-enable opaque types as long as they remain
opaque until erasure, but they must be erased to their underlying
type - so an efficient and correct representation should be found.