Use correct tree for child unapply context
Prevents `ClassCastException` in some GADT scenarios
and makes valid code compile in others.

The parent context tree could be the case def itself.
In that case creating a child context with the same tree
breaks the logic to restore saved type bounds in `typedCase`,
because `enclosingCaseDef` returns the child instead of the parent.
This leads to saving modified bounds in the child context
but later trying and failing to restore them in the parent context.

The leaked bounds later cause type errors in the best case
or propagate incorrect information to erasure in the worst case.

Using the tree of the unapply itself for the child context
means that `enclosingCaseDef` will return the parent context.