Travis CI: make a command error fatal
I saw a CI log where the failure to build and bootstrap didn't kill the
job, trying to fetch that starr version to testOsgi did...  Previously
this wasn't the case because there was a "set -e" (aka "-o errexit")
present which I just recently removed, because I confused it with "set
-x" (which is "-o xtrace"). -.-

There's a bit of an holy war on "set -e" usage but I (still?) tend to
lean on the "don't rely on it" side.  I'd initially gone with my "append
everything important with `|| travis_terminate 1`" solution, but I found
the "make it one big command with `&&`" way to be perhaps better because
I think it's harder to not do the wrong thing that way (i.e.  forget to
append travis_terminate).  I think the downside is a failure reports the
entire chain of commands rather than explicitly which one failed...

However, as many other scripts use "set -e" so we'll keep it consistent.