Patterns under irrefutable extractors are fixed
The reason that:

    sealed trait May[+A]
    final case class Just[+A](value: A) extends May[A]
    final case class Nada()             extends May[Nothing]

    object Possibly {
      def unapply[A](may: May[A]): Some[May[A]] = Some(may)
    }

    def usingMay[A](may: May[A]) = may match { // missing Nada case
      case Possibly(Just(a)) => a
    }

... also reports "Just" in addition to "Nada" is that the compiler
doesn't know under what conditions the Possibly extractor returns Nada
(which is the root of the unhandled case) so it gives you both the
input's possible types: Just and Nada.  Like before (with the "are
custom extractors idempotent") I think `Some[may.type]` fixes that, but
noone writes that.