Fix compiler crosstalk w/ an empty package reset
The point of using a "resident" compiler is to hold on to a single
Global which initialises a bunch of expensive things, so you save that
cost when you execute multiple compile Run's on various code.  One of
those things is constructing and populating various data structures
based on the runtime that the compiler finds itself executing on, the
standard library and JDK, for instance.  For instance constructing the
Symbol's and Scope's for various important parts of the classpath.

Wiping out all of that would (I predict - the mechanics aren't there and
I didn't get far enough to build and test it) negate the intent of
reusing a Global.  So instead I just wipe out the empty package, which
works well because production code doesn't (shouldn't) live there but
test code frequently does.  Fortunately it's logic that lives in
testkit-land so it's imperfections don't cascade into production usage.