Move & truncate the MiMa filters
First of all, move the damn MiMa filters outside of build.sbt!  In
build.sbt it overpowers the rest of the build definition and also (it
would seem) IntelliJ is better at handling Scala files than sbt files.

Now, if a contributor needs to add filters, they can be advised to
change "project/MimaFilters.scala".

Secondly, bump the `mimaReferenceVersion` (which is used to define
`mimaPreviousArtifacts`).  Normal projects (without forwards
compatibility) want to always bump that, so that new methods don't go
missing.  But for Scala it's less of a problem.  However, it is because
it means we have an always appended list of exclusions, and those might
end up hiding real problems.

While I was at it I fixed `mimaReportBinaryIssues` it can be called
directly, by resetting `ThisBuild / mimaFailOnNoPrevious`.