Fix range position for vars with default initializer
Pass the position of the RHS explicitly rather than
using `rhs.pos` within TreeBuilder, as that doesn't
carry the position of the `_` represented by the
singleton tree `EmptyTree`.

Change the quasiquote parser to also pick up the change.

Also, tweaks in DefinitionConstructionProps that came out of fixing the
quasiquote parser change.

Co-authored-by: Dale Wijnand <dale.wijnand@gmail.com>