improvement on definition of inits?
Hello,

I am just speculating (wildly and gratuitously, no doubt) as to whether it could be desirable at all to change the documentation of inits by taking into account both the Haskell definition, and the fact that inits can be defined recursively.

Haskell definition of inits: https://hackage.haskell.org/package/base-4.14.1.0/docs/Data-List.html#v:inits

Recursive definition of inits (in Haskell): https://www2.slideshare.net/pjschwarz/the-functional-programming-triad-of-folding-scanning-and-iteration-a-first-example-in-scala-and-haskell-polyglot-fp-for-fun-and-profit#18

I think in a way the current Scala documentation for inits is quite good because it depends on the definition of init, and so it is very simple and brief. So in a way there is probably no need to change it.

I realise I may be opening a small can of works, e.g. if the documentation for inits were to change, what about that for tails?

The text that I have replaced the original with is just a very first attempt. e.g. the new definition and the sample code could use Seq instead of List. Also, should this whole idea end up being judged to be of any merit, I can imagine the text having to undergo a lot of changes and tweaks to get it to conform to maintainers' notions of what is acceptable in Scala docs, and what is customary, what is superfluous, etc.

I would completely understand if nothing comes of this. I just thought I ought to ask.

Thanks in advance for taking the time to read this.

P.S. https://scastie.scala-lang.org/pmIfdCKrQP2FrNoxlX3xJw