Restore two blank lines bailing the REPL
LineReaderImpl handles EOFError (which is a subclass of SyntaxError) as
needing to read more input.  So in order for our non-JLine-specific "You
typed two blank lines. Starting a new command." behaviour in ILoop to
function we need to bail (with a regular SyntaxError) if we read two
newlines.

Perhaps we should also bail if we fail to parse (Error) but I left that
as is, to avoid breaking something.