Fix mimaReportBinaryIssues, not run aggregate
When you run a task in the shell, it uses the current project (normally
root) and runs the tasks on the project and the project it aggregates.
That's why "compile" compiles all and "test" tests all.  But "testAll"
is our custom thing, defined only on the root project, which refers to
specific tasks to run.  So while just "mimaReportBinaryIssues" does the
right thing in the shell, it only checks root (which is a noop) when
used in "testAll".  So back to specifying both - but at least I still
fixed unqualified "mimaReportBinaryIssues" from working in the shell.