Fix regression in specialization of genericWrapArray(Array(..))
The extension to the optimization in Cleanup of Array.apply
assumed that the argument array was non-primitive.  This broke
compilation of spire (code minimized to test/files/run/sd760a.scala)

I also found a path to the same bug without specialization, by using
a generic type parmater bounded by a primitive. This is in
test/files/run/sd760b.scala

The fix here is straight forward -- disable this optimization when the
argument array is primitive.