Cleanups from fixing a bug in typedConstructorPattern
"sym in ${sym.owner} is a java8+ default method." comes up a few times,
making it a noisy log message.  I considered dropping it but I just
lowered it for now.

Removed a reference to "parameterized extractors" which I think I
understand but we never got/implemented.  Perhaps there are
simplifications that can be made based on them not existing and no plans
for them to exist, but I don't know enough to know where to look.