Make restarr <-> reload switching faster
Define the `ThisBuild / target` key and rewire the build to it so the
`restarr` command can redefine it and `ThisBuild / buildDirectory`,
which is also used.  This allows switching between both versions much,
much faster.

When you make any source change you'll need to do the more
time-consuming `restarrFull` but the point is that switching back to the
previous starr (via `reload`) will be fast because the build products
haven't been wiped out in the process.

There's a small amount of projects that use "target" as their project
base, which is a harder value to change (requires a `reload` during
`restarr` which wipes out the session settings, so not 100% clear how to
even do it).  Looking at them I think it might be fine and a small smoke
test seems to show it is fine.