Avoid classfile parsing of specialized variants just to unlink them
Since 8ae0fdab, the specializion phase eagerly info transforms
all of FunctionN and TupleN. This was done to let us turn off
needless specialization info transforms (which incurs classfile parsing
up the base classes looking for @specialized annotations) of types after
the specialization tree transform is done.

However, in combination with an old fix for scala/bug#5545, we end
up parsing all of the class files of all the variants, just to
unlink them in favour of the info-transformed types.

I note that the test for scala/bug#5545 no longer crashes if the fix
is removed. I have have not investigated the reason.

This commit reworks the scala/bug#5545 to just unlink the stale
symbols directly, rather than calling `.info` to parse them and
do the same after noticing the ScalaRaw attribute.