Perform override checks again for two Java-defined methods
The shortcut that's removed here was added in PR 8643 but turned out
to be incorrect. The bug fixed by that PR remains fixed, because the
shortcut was only added as a "second layer", the underlying bug was
fixed as well.

Note that RefChecks only runs on Scala classes, so we only perform
override checks of two Java-defined members if a Scala class inherits
them. These checks are not optional, as shown by the test cases added
here.