Handle Singleton types in patmat's outer prefix align testing
Singleton abstract types are isStable, but they don't have a term
symbol.  But, at least in the given test case, they do have a type
symbol (the type parameter symbol) so we can construct a fresh singleton
type using that type symbol and use that to determine that no outer test
is needed for the prefix.  That hinges on the assumption that the
machinery around Singleton is successfully enforcing that it all ends up
deriving from a single value and two types are never from different
prefixes.

Alternatively we can just use `pre.typeSymbol == NoSymbol` as a guard
and always emit an outer test on the prefix.  That undermines that
Singleton abstract types are stable, but I can't tell in the context of
outer tests whether that's the right or the wrong choice...