Override checks between Java-defined members only for mixins
Override checking between Java-defined members can and should be
skipped: javac will do the checks anyway, and applying Scala's rules
can lead to false errors.

However, when mixing in a Java interface into a Scala class, default
methods in the interface have to be checked according to Scala's
rules, because Scala linearization applies in this case.