Speed up BSP import of project into IDE by skipping JMH source gen
JMH has source generators that wrap the user-written benchmarks in
synthetic code that wraps it will timing loops.

By default, BSP import runs all source generators. This is useful
for things like protobuf bindings, where the user wants to code
against the generated APIs.

But the JMH generated code isn't useful to see in the IDE and
by skipping it we can import the project without needing to
trigger compilation of scala-library / bench.