Tweak test to wait for expiring thread
The test for exception handling waits for thread
to die, perhaps to ensure all actions are complete.
Previously, throwing terminated the thread, but in
JDK 17, the thread's exception handler is the pool's.
The test is updated to wait some long period for
the thread to expire of natural causes. Wonder why
the replicants in Blade Runner didn't change their
thinking from "retirement" and "termination" to
"Hey, I've got a pretty healthy keepAlive."