Configure OSGI bnd tool for reproducible builds
We already set use the Git timestamp for the OSGi version number,
but we were still getting diffs in the META-INF/MANIFEST.MF
file in the `Bnd-LastModified` header.

This commit sets the REPRODUCIBLE flag to tell the OGSi bundler tool
to omit that header and ensure that the current timestamp does not
leak into other Zip entry metadata.

Together with the upgrade to SBT 1.5.x this seems to make the
resulting JARs successively builds from a given commit identical.

There have been some behaviour changes in bnd that affect our
output. These seem to be valid bug fixes, but I have opted to
stay with our status quo with a little extra configuration.