No ClassTag for matching abstract type tests
When the static type matches the type test, there's no need for a
ClassTag.  So it shouldn't emit an unchecked warning (checkCheckable).

Also, in that case, even if a ClassTag is available, there's no need to
use one (extractorForUncheckedType).

Compared to previously, two additional things will change: if the value
isn't of its static type (e.g. a cast was used) or it's _null_,
previously this would fail early with the case not matching.  Now,
instead, the case will match and (likely) it will fail as soon as the
value is used.