Don't expand the name of accessors
I couldn't find a single use case where this is needed.
The callee is usually either not an accessor at all
or all declarations in scope are looped over and expanded.

On the other hand this fixes a bug in specialization.
After all the accessor(s) could be implementing an abstract method.