Prevent compilation of library/reflect/compiler on bsp import
Importing / refreshing the BSP triggers compilation of the compiler
(and library, reflect).

It's rooted in the `buildTarget/resources` bsp call, which runs
`bspBuildTargetResources file:/Users/luc/scala/scala13/#scaladoc/Compile`.
The issue can be reproduced by just calling `scaladoc/resources` in sbt.

The fix is to only look at the `externalDependencyClasspath`
(not the `dependencyClasspath`) when searching for webjars as scaladoc
resources.