Restore partest crash handler
The unused crash handler was refactored out by paulp
in current partest's origin story, and not by more recent
refactorings. A previous tweak to handle Crash as a
TestState did not show or log a stack trace, so the
log would read: "error: null", for example.

The crash handler writes the stack trace to the log.

If there is a check file and the exception is a FatalError
from the compiler, then only the exception message is
logged, so that a neg test can diff it reliably.
If the log file matches the check file, the test passes.
The output will say "fatal error:" to help distinguish
from other errors.

The display of test results in AbstractRunner need not
provide additional output about the crash. The terse
mode icon for a Crash is changed from X to ?.

There is no distinction between an expected Crash
(as DNC or does not compile) and an unexpected error.