Reduce default capacity of classpath's pool of ZipFile for each JAR
On systems with large values of "available CPUs", highly concurrent
compilation, and large classpaths, the pool was observed to have
exhausted file handles.

Reducing the pool to 4 doesn't actually incur locking if 5+
concurrent compilers access a given JAR, but it will require
all but the first 4 to open/close the file and ZipFile.