Lift try blocks to methods in value class invocations
In `def t = (try c catch { .. }).meth`, it looks like the stack
is empty when loading `c`. But if `c` is a value class, the call is
translated to `C.meth$extension(try c catch { .. })`, so the
`try` expression needs to be lifted to a local method.

Make UnCurry aware of this.