Allow ? as a wildcard even without -Xsource:3
Like https://github.com/scala/scala/pull/9721, the goal here is to ease
cross-compilation between Scala 2 and 3 code as we prepare for `_` as wildcard
to be deprecated in Scala 3 (cf https://github.com/lampepfl/dotty/issues/5379).

This is technically a source-breaking change, but using a type named ? without
backticks is deprecated since 2.13.6 (#9626) and defining a type named ? without
backticks is disallowed since 2.13.7 (#9735). The only known user of ? as a type
name was the kind-projector plugin, but support for this syntax was removed from
the plugin a year ago (https://github.com/typelevel/kind-projector/pull/186),
so I believe this change won't be disruptive.